requirements.txt
aiohttp
beautifulsoup4
googlesearch-python
pyyaml